Oops j/k it was passing tests because there were no tests. But fixed a superficial bug and added tests and now it's good
git-svn-id: file:///home/runner/work/tom7misc-sync-action/tom7misc-sync-action/upstream/trunk@4801 c7e11af0-4c7a-11f1-9edd-117cb6f09d55

+++ grad/expression.h
@@ -98,7 +98,7 @@ struct Exp {
     const Exp *PlusV(const std::vector<const Exp *> &v) {
       CHECK(!v.empty());
       const Exp *r = v[0];
-      for (int i = 1; i < v.size(); i++) {
+      for (int i = 1; i < (int)v.size(); i++) {
         r = PlusE(r, v[i]);
       }
       return r;
